Overhaul event states
This is a breaking change. Events can now be either pending, triggered,
or aborted. Event::trigger triggers the event. This schedules the event
to be fired immediately. Event::abort aborts the event. This calls the
Event::Aborted method. To check the state of an Event, use
Event::is_pending, Event::is_triggered, and Event::is_aborted.